+++ ddmrp/models/procurement_group.py
@@ -1,8 +1,7 @@
-# Copyright 2016-18 Eficent Business and IT Consulting Services S.L.
-#   (http://www.eficent.com)
+# Copyright 2016-20 ForgeFlow S.L. (http://www.forgeflow.com)
 # Copyright 2016 Aleph Objects, Inc. (https://www.alephobjects.com/)
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 # Copyright 2018 Camptocamp SA https://www.camptocamp.com
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
 from odoo import api, models
 
@@ -16,3 +15,31 @@ class ProcurementGroup(models.Model):
         automatically procure from orderpoints and to automatically
         reserve stock moves."""
         return True
+
+    # UOM: (stock_orderpoint_uom):
+    @api.model
+    def run(self, procurements):
+        Proc = self.env["procurement.group"].Procurement
+        indexes_to_pop = []
+        new_procs = []
+        for i, procurement in enumerate(procurements):
+            if "buffer_id" in procurement.values:
+                buffer = procurement.values.get("buffer_id")
+                if buffer.procure_uom_id and \
+                        procurement.product_uom != buffer.procure_uom_id:
+                    new_product_qty = procurement.product_uom._compute_quantity(
+                        procurement.product_qty, buffer.procure_uom_id)
+                    new_product_uom = buffer.procure_uom_id
+                    new_procs.append(Proc(
+                        procurement.product_id, new_product_qty, new_product_uom,
+                        procurement.location_id, procurement.name,
+                        procurement.origin, procurement.company_id, procurement.values
+                    ))
+                    indexes_to_pop.append(i)
+        if new_procs:
+            indexes_to_pop.reverse()
+            for index in indexes_to_pop:
+                procurements.pop(index)
+            procurements.extend(new_procs)
+
+        return super(ProcurementGroup, self).run(procurements)