+++ ddmrp/wizards/multi_level_mrp.py
@@ -1,5 +1,5 @@
-# Copyright 2018 Eficent Business and IT Consulting Services S.L.
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+# Copyright 2018 ForgeFlow S.L. (http://www.forgeflow.com)
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
 from odoo import api, models
 
@@ -10,11 +10,10 @@ class MultiLevelMrp(models.TransientModel):
     @api.model
     def _exclude_from_mrp(self, product, mrp_area):
         """ Exclude from MRP scheduler products that are buffered. """
-        res = super(MultiLevelMrp, self)._exclude_from_mrp(product, mrp_area)
-        ddmrp = self.env['stock.warehouse.orderpoint'].search([
+        res = super()._exclude_from_mrp(product, mrp_area)
+        ddmrp = self.env['stock.buffer'].search([
             ('location_id', '=', mrp_area.location_id.id),  # child of?
             ('product_id', '=', product.id),
-            ('buffer_profile_id', '!=', False),
         ])
         if ddmrp and not self.env.context.get("mrp_explosion"):
             return True